requirements.txt
aws-cdk.core
aws-cdk.aws-ec2
aws-cdk.aws-codebuild
aws-cdk.aws-ecr
aws-cdk.aws-ecs
aws-cdk.custom-resources
jsii
-e .
